--- src/main/kotlin/com/unrec/imdb/search/model/AkasRecord.kt
@@ -1,12 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class AkasRecord(
-    var titleId: String = "",
-    var ordering: String = "",
-    var title: String = "",
-    var region: String = "",
-    var language: String = "",
-    var types: String = "",
-    var attributes: String = "",
-    var isOriginalTitle: String = "",
-)
