--- src/main/kotlin/com/unrec/imdb/search/model/BasicsRecord.kt
@@ -1,13 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class BasicsRecord(
-    var tconst: String = "",
-    var titleType: String = "",
-    var primaryTitle: String = "",
-    var originalTitle: String = "",
-    var isAdult: String = "",
-    var startYear: String = "",
-    var endYear: String = "",
-    var runtimeMinutes: String = "",
-    var genres: String = "",
-)
