--- src/main/kotlin/com/unrec/imdb/search/model/CrewRecord.kt
@@ -1,7 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class CrewRecord(
-    var tconst: String = "",
-    var directors: String = "",
-    var writers: String = ""
-)
