--- src/main/kotlin/com/unrec/imdb/search/model/EpisodeRecord.kt
@@ -1,8 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class EpisodeRecord(
-    var tconst: String = "",
-    var parentTconst: String = "",
-    var seasonNumber: String = "",
-    var episodeNumber: String = "",
-)
