--- src/main/kotlin/com/unrec/imdb/search/model/NameBasicsRecord.kt
@@ -1,10 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class NameBasicsRecord(
-    var nconst: String = "",
-    var primaryName: String = "",
-    var birthYear: String = "",
-    var deathYear: String = "",
-    var primaryProfession: String = "",
-    var knownForTitles: String = "",
-)
