--- src/main/kotlin/com/unrec/imdb/search/model/PrincipalsRecord.kt
@@ -1,10 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class PrincipalsRecord(
-    var tconst: String = "",
-    var ordering: String = "",
-    var nconst: String = "",
-    var category: String = "",
-    var job: String = "",
-    var characters: String = "",
-)
