--- src/main/kotlin/com/unrec/imdb/search/model/RatingsRecord.kt
@@ -1,7 +1,0 @@
-package com.unrec.imdb.search.model
-
-data class RatingsRecord(
-    var tconst: String = "",
-    var averageRating: String = "",
-    var numVotes: String = "",
-)
