+++ src/main/kotlin/com/unrec/imdb/search/record/Record.kt
@@ -1,0 +1,7 @@
+package com.unrec.imdb.search.record
+
+import com.unrec.imdb.search.entity.Entity
+
+interface Record {
+    fun toEntity(): Entity
+}
